Refactor Draftail exports to match exposed module
